Implement substitution rules.
Note that our substitution rules also have implications for usages of
type variables even *before* substitution. For example, CF considers
`@NullnessUnspecified T` to always "have unspecified nullness," and thus
it considers it safe to dereference in "lenient mode." In contrast, we
consider "unspecified nullness" to be only the *lower* bound, so we will
want to forbid dereferences (as it may be instantiated with a nullable
type argument).

Full handling of these implications will require additional changes,
likely including changes to CF itself.