Default unbounded wildcards' upper bounds to @Nullable.
This requires an API currently available only in our local copy of the
Checker Framework.

(Note that this whole project already requires our local copy of the
Checker Framework in order to work correctly. This particular commit is
just the first to introduce a "linkage" dependency on that local copy.
That is, this commit will actually prevent the project from *compiling*,
rather than letting it compile but then behave incorrectly.)