Handle enums and literals with a TreeAnnotator.
The enum change eliminates the need for the ENUM_CONSTANT API we have in
our CF fork.

The literal change doesn't really accomplish anything specific, but:

- It seems slightly nice to handle enums and literals in the same way.

- I generally like the idea of moving away from encoding rules in our
package-private annotations, since it might be nice to eliminate those
annotations someday (notably, to ensure that they aren't written to
bytecode), as discussed in a TODO in NullSpecAnnotatedTypeFactory.

That said, I haven't fully eliminated the need for @DefaultFor, and I
ran into some trouble when I briefly attempted to.