Switch to NoElementQualifierHierarchy, and override createQualifierKindHierarchy.
Then, eliminate our package-private @Nullable and @NullnessUnspecified.
They are no longer necessary because we have eliminated the final piece
of declarative configuration.

Another nice aspect of eliminating @SubtypeOf is that, thanks to our
weird subtyping around "unspecified nullness," those annotations were
presenting an incomplete picture. To be fair, we're now replacing that
with an identical incomplete picture in createQualifierKindHierarchy.
But at least that gives us a good place to document things. And that
incompleteness remains fairly well contained: The closest thing to a
"leak" is that we still don't consider "unspecified nullness" to be a
top.

Also, eliminate our leastUpperBound and greatestLowerBound
implementations. We get them for free from NoElementQualifierHierarchy.
I should have been using it instead of ElementQualifierHierarchy from
the beginning. Oops.