Recognize null checks on type-variable usages.
This commit is perhaps surprisingly large. That's because it requires us
to support a 4th "kind" of type-variable usage: In addition to
supporting a usage with each of the 3 additional-nullness values, we
must also support a usage that has been refined by dataflow to be
non-nullable, even if the corresponding type parameter permits nullable
type arguments.

The natural way to do this is to apply @NonNull to such non-nullable
usages. Yes, I had previously replaced @NonNull with
@NoAdditionalNullness. Sorry, Werner -- now I understand :) I will
rename it back in a subsequent commit, since the diffs from this one are
already large.

But conceptually, we still want for most null-aware type-variable usages
to be have "no additional nullness." This means that we can't keep
applying @NoAdditionalNullness to them by default, since that annotation
now really means "@NonNull."

But... we still need to apply defaults to type-variable usages in
*non-null-aware* code! To accomplish this, we apply a default in *all*
cases but then subsequently remove it @NonNull whenever it appears on a
type-variable usage.

Then we fiddle with a couple other things, and it works!