Do not write computed annotations to bytecode.
We don't want for consumers to depend on them.

Plus, I had a weird type.invalid.conflicting.annos error when testing
against Truth. (I was testing with -AaliasCFannos and -AcheckImpl but
not -Astrict. I believe that at least the first 2 flags were necessary
to reproduce the problem.)

core/src/main/java/com/google/common/truth/TruthFailureSubject.java:155: error: [type.invalid.conflicting.annos] invalid type: conflicting annotations [@Nullable, @NullnessUnspecified] in type "@Nullable @NullnessUnspecified String"
    String value = factsWithName.get(firstNonNull(index, 0)).value;
                                                            ^

The @NullnessUnspecified annotation was being added by the "write
annotations back to Elements/bytecode" path.

Before this commit:

  final java.lang.String value;
    descriptor: Ljava/lang/String;
    flags: (0x0010) ACC_FINAL
    RuntimeVisibleAnnotations:
      0: #41()
        org.checkerframework.checker.nullness.qual.Nullable
    RuntimeInvisibleTypeAnnotations:
      0: #38(): FIELD
        org.jspecify.annotations.NullnessUnspecified

After this commit:

  final java.lang.String value;
    descriptor: Ljava/lang/String;
    flags: (0x0010) ACC_FINAL
    RuntimeVisibleAnnotations:
      0: #39()
        org.checkerframework.checker.nullness.qual.Nullable

This is *maybe* a CF bug: It presumably should not write
@NullnessUnspecified (or even compute it in the first place), since
@Nullable is already present. However, we are doing a combination of
many weird things. Possibly of particular relevance here: We are making
the CF @Nullable an alias for ours, *and* we are using the CF @Nullable
as a *declaration* annotation (rather than a type annotation), since we
are compiling with `-source 7 -target 7`. I'm not even sure that this is
supported by CF, and even if it is, I can't imagine that it's an urgent
problem for them to fix.

Thankfully, not writing computed annotations to bytecode is something
that we want to do anyway. The fact that it works around this
self-inflicted problem is a terrific bonus.

(And of course we are doing weird things in general, such as with our
conditional defaulting rules and non-transitive subtyping rules. Those
are probably not specifically relevant here, but you never know.)

I did also take this opportunity to investigate my earlier fears about
not writing computed annotations to bytecode. (Those fears were largely
allayed by the discovery that postProcessClassTree exists for exactly
this purpose, but I figured I'd check while I had the chance.) I looked
at one particular case, the case of a "? super Foo" wildcard in
Correspondence.java, specifically in the nested class
com.google.common.truth.Correspondence$Transforming. I wanted to make
sure that CF was not only omitting the computed annotations but also not
writing any sort of "extra" bound to the bytecode. It doesn't appear to
be writing any such bound, though I admit that (as noted in an earlier
implementation comment) I'm not sure there's actually a way to write "an
extra bound" to bytecode, only a way to write an annotation on a type
parameter or wildcard.

Before this commit:

  private final com.google.common.base.Function<? super A, ?> actualTransform;
    descriptor: Lcom/google/common/base/Function;
    flags: (0x0012) ACC_PRIVATE, ACC_FINAL
    Signature: #13                          // Lcom/google/common/base/Function<-TA;*>;
    RuntimeInvisibleTypeAnnotations:
      0: #15(): FIELD, location=[TYPE_ARGUMENT(1), WILDCARD]
        org.jspecify.annotations.NullnessUnspecified
      1: #16(): FIELD, location=[TYPE_ARGUMENT(1)]
        com.google.jspecify.nullness.NonNull
      2: #15(): FIELD, location=[TYPE_ARGUMENT(0), WILDCARD]
        org.jspecify.annotations.NullnessUnspecified
      3: #17(): FIELD, location=[TYPE_ARGUMENT(0)]
        org.jspecify.annotations.Nullable
      4: #15(): FIELD
        org.jspecify.annotations.NullnessUnspecified

After this commit:

  private final com.google.common.base.Function<? super A, ?> actualTransform;
    descriptor: Lcom/google/common/base/Function;
    flags: (0x0012) ACC_PRIVATE, ACC_FINAL
    Signature: #13                          // Lcom/google/common/base/Function<-TA;*>;