Make getUpperBounds call AnnotatedTypeMirror methods directly.
I don't expect this to change any results, but it feels a bit clearer --
especially for the wildcard case, in which there's a distinction between
looking only at the bounds of the wildcard and looking at a combination
of those and the bounds on the corresponding type parameter.

I noticed this because of recent changes to the
AnnotatedIntersectionType API. Those changes exposed a getBounds()
method -- which we now use, enabling us to call get*Bound* methods in
all 3 cases of the switch().