Fix defaulting logic by using new hooks.
The new hooks are in
https://github.com/jspecify/checker-framework/commit/ddec8ff2f0a68d30f31509a4ad24f6c48dbaeb03

This commit doesn't fix any errors with our existing samples. However,
it prevents problems that would arise in a forthcoming sample (which is
not yet released because it requires additional fixes in this checker
:)). And it eliminates the incorrect defaults I'd see transiently during
debugging at times.

Some details:

- Set the "default default" with per-element defaults rather than with
checkedCodeDefaults.

- Set per-element defaults with my new `populateNewDefaults` hook. My
old approach did not always work, as it relied on CF's calling
`annotate(foo)` before `annotate(childOfFoo)`, and CF does not guarantee
that.

- Eliminate the old hack of "annotate type-variable usages with
`nonNull`, then un-annotate them." I think this was conflicting with my
forthcoming attempt to actually annotate some type-variable usages as
`nonNull`. (As a workaround, I had started to put into place logic to
remove only `nonNull` annotations _that were not present
beforehand_. But that was getting ugly. I'm not sure whether I had
it all working or not, but with the change to the "default default"
logic, it's no longer necessary.)

- Use my modified CF's logic for inheriting defaults. I have removed
parts that inherited from parent packages.