Allow any type argument for AtomicReference if the caller provides an initial value.
This is how it should have been in the first place. Oops.

This also shows another reason that checkTypeArguments is not the right
tool for this job: The type arguments themselves are fine (especially in
"non-execution" contexts like `private final AtomicReference<Foo>;`);
the problem is the lack of an argument when invoking the constructor
with certain type arguments.