Handle the case of `T` vs. `@NonNull T`.
I'm not completely clear on whether this "should" arise at all, but it
certainly arises a lot in Guava, so let's handle it.

This also appears to fix handling of "self types" (`class Foo<F extends
Foo<F>>`). I had assumed that the fix for that would be much more
invasive, but it turns out to be this easy.