Look for defaulting annotations in stub files.
This fixes a huge number of problems I saw when testing against Guava in
strict mode.

However, a huge number of problems remain. The cause? Sometimes
`getDeclAnnotations` returns an empty set even when annotations are
present on the stub file.

This appears to happen when a stub file is first processed during the
processing of *another* stub file.

It seems that `AnnotatedTypeFactory.getDeclAnnotations()` tries to avoid
this problem by never caching any results that it produces during
stub-file parsing:

https://github.com/typetools/checker-framework/blob/ac8e4c356c947f89104e7806e811cf0129eeb5a5/framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java#L3370,L3384

However, the problem is deeper than that: It's not that we're seeing
caching of _the set of annotations_. Rather, we're seeing caching of
_defaults_.

Here's an example: During the parsing of the stub file for `String`, we
need to refer to `Object`. As part of that, we go to compute types for
APIs in `Object`. As part of _that_, we go to apply _defaults_ to those
APIs. At that point, we call back into `getDeclAnnotations`. But we're
in the middle of parsing (specifically, parsing the stub file for
`String`), so `getDeclAnnotations` doesn't consult `stubTypes`. Thus, it
returns an empty set of annotations for `Object`. And thus, we apply
defaults to `Object` as if there is no `@NullAware` annotation on the
type. Then we cache those defaults. So when we come back later to ask
about `Object` specifically, the damage has already been done.

Where does this stub-file recursion arise?

I've seen it in `addAnnotationsFromDefaultForType`. I tried making that
a no-op, since I don't think we need it.

Then I saw it in `postDirectSuperTypes`, which calls
`addComputedTypeAnnotations`. I tried commenting that section out. I
haven't yet seen any evidence that we need it.

So a possible way forward is to cut off those 2 sources of recusion and
hope for the best.

Another possible way forward is to remove the `isParsing()` check in
`AnnotatedTypeFactory.getDeclAnnotations()` and and the `parsing` check
in `StubTypes.getDeclAnnotation()`. That is, we would _never_ accept
"I'm currently parsing, so I'm going to say there are no annotations
here" as an answer.

Naturally, this triggers infinite recursion. I was able to work around
that by turning 2 methods in `DefaultQualifierForUseTypeAnnotator` into
no-ops, `getSupportAnnosFromDefaultQualifierForUses` and
`getHierarchiesNoDefault`. I don't believe we need these, and that seems
to work.

(Note that merely turning those 2 methods into no-ops but preserving the
`isParsing()` and `parsing` checks does _not_ solve the problem.
Evidently there is still some defaulting that happens during recursive
parsing, so it still receives empty sets of annotations. A widespread
need for defaulting isn't too surprising, so I haven't dug into it
further. I think I have spent enough time on this :))

I'm not entirely sure which direction I'll approach this from yet. I
suspect I'll take the second approach (removing `isParsing()`+`parsing`
checks and avoiding recursion in `DefaultQualifierForUseTypeAnnotator`).
It feels like a step toward _more_ honesty: `getDeclAnnotations` will
always return the right annotations (or at least overflow the stack if
it turns out that it can't), and I don't have to remove code that I
don't fully understand in `postDirectSuperTypes` (which I fear would
lead to more "lies" in the future).

Or maybe I'll think of something else. Certainly all of this feels a bit
fragile. It's not hard to imagine that some other source of recursion
might arise in the future.

TODO(cpovirk): File a bug against upstream CF about this.