Make `<E extends Enum<E>>` a warning instead of an error.
I should probably look more at this someday, but:

In short, the problem seems to be that I'm not implementing type
equality in the way that the spec says. The fix is to do that... mostly.

It turns out that the method that I'm implementing may or may not be
intended to implement type equality per se.

Anyway, the sort of error that this fixes is as follows (rendered with
extra-verbose types that include `!` for projections to non-null):

```
check_api/src/main/java/com/google/errorprone/ErrorProneFlags.java:94: error: [nullness] incompatible type argument for type parameter E of Enum.
  public <T extends Enum<T>> Optional<T> getEnum(String key, Class<T> clazz) {
                         ^
  found   : T*
  required: Enum<T!>!
```

For some reason, this problem does _not_ appear if the self-typed class
is declared in source code. I haven't checked whether the problem can
appear with any bytecode class or only one that appears in stub files.

Again, I should look at this more to understand it better.