Simplify visitTypevarSupertype special case.
I'm not sure at what point this simplification became possible. I had
suspected my new
https://github.com/jspecify/checker-framework/commit/fd655e220042b3288cc407a447e9eff68ddc1a06,
but it looks like the simplification was possible even without that.

I'm still not sure why this case differs from the subtype case, nor why
it differs from the wildcard-supertype case. (Perhaps it's related to my
override of visitTypevar_Typevar?)

It's entirely possible that this special case is in fact still necessary
but that I just don't have sample inputs that demonstrate that. (Maybe
that was even always the case; I don't remember if I ever had a failing
sample to motivate the specific special handling that I'm removing.)