Eliminate custom determinism check.
It was no longer excluding anything in particular.

But even beyond that, "our own sort of [alleged] determinism check" was
probably not doing what we'd want: I think it ignored subexpressions.
So, even if there _were_ some kind of expression that our check
concluded was nondeterministic, we'd currently still consider an
expression that _contained_ that expression to be deterministic. I
think? This would need more thought if we wanted to go down that road.
Maybe we'd want to hook into the existing CF determinism check, or maybe
we'd want to visit all subexpressions as part of our own check.