Start phasing out the phrase "code not nullness aware."
I forget what prompted me to do this right now, but it's going to happen
eventually:

- Whatever term we choose for our "in the club" annotation, it's not
going to be "NullAware."

- The concept here is not merely "The type appears in code that is not
annotated `@NullAware`" but all that _plus_ "and it doesn't have a
`@Nullable` annotation on it."

- The name is also wrong in the opposite direction, at least
potentially: If we have a `@NullnessUnspecified` type annotation, then
it will be possible to have a type in `@NullAware` code with an
additional nullness / nullness operator of `CODE_NOT_NULLNESS_AWARE`.

- Maybe the trigger for doing this now was seeing
`CODE_NOT_NULLNESS_AWARE` in code that was itself `@NullAware`, since it
depended on code that was not `@NullAware`?

I have also squeezed in a few references to "nullness operator," the
term that may replace "additional nullness" when we revise the subtyping
rules to contain "project to non-null."