Make map.get after containsKey return "just V," even if V is null-inclusive.
Maybe also fix `refine` when one of the types is an annotated
type-variable usage and the other is an unannotated type-variable usage?
But if that change was necessary, then we probably have deeper problems
when CF calls methods like CFAbstractValue.mostSpecific internally.
Hopefully it doesn't come up much. If it does, we'll probably need to
provide our own implementation of CFAbstractValue.