Work around ImmutableMultiset problem in a different way.
It turns out that I was wrong about addDefaultAnnotations in my previous
commit message: addDefaultAnnotations is called when we need it after
all.

I got confused because I had implemented it to look for an
AnnotatedNullType instead of an AnnotatedTypeVariable whose *bound* is
an AnnotatedNullType. (The method is not called on every component of
the type, only the top-level type -- unless I'm confused again :))