Make checkArgument(x != null) guarantee that x is not null.
This is the same thing that
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/0795eb8d09c66276ba48c41209c07fa04ed7d860
did for checkState.

I have somewhat mixed feelings about this: I don't want to encourage
people to write `checkArgument(arg != null)` if they find that
`checkNotNull(arg)` produces an error (which happens if they've
annotated `arg` as `@Nullable`). But hopefully our suggestion of
`requireNonNull(arg)` will be more palatable, anyway. And hopefully this
doesn't come up *too* often (though I know I've seen complaints about
the signature of checkNotNull before, so clearly it comes up enough).

On the other hand, I saw cases like the following:

```
Foo associatedValue = someMap.get(arg);
checkArgument(associatedValue != null, "Unrecognized arg %s", arg);
```

And checkArgument feels exactly right there: Much like with the
checkState case, it feels clear to me that the author of such code knows
that the check may fail (if and only if the caller has done something
wrong). Thus, I don't feel that we're sweeping an issue under the rug by
treating the null check as legitimate.

Additionally, if we do want to issue an error for such checks -- at
least for the `checkArgument(arg != null)` case -- then the right place
to report it seem to me to be the checkArgument call itself. After all,
what we'd want to report is "This checkArgument check may fail," not
"This later dereference may fail," since we would never get to the later
dereference because of the checkArgument failure! So even if we go that
direction, I'd still like to make this current change, just
*supplemented* by the additional validation of the checkArgument call
itself.