Use type.getTypeVariable when wildcardToTypeParam fails.
This fixes a regression from
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/9348e76115088149d55d9495a5c720a5f1d6cd34
on code like the following:

```
Class<?> c = Class.forName(name);
```

In that case, visitWildcard was replacing the variable's type with `Class<?
extends Object>` but leaving the method's return type as `Class<?>`. Then it
was finding those types to be incompatible.

The difference seems to be that wildcardToTypeParam fails for types read from
bytecode.

In addition to using type.getTypeVariable in visitWildcard, I've decided to
also use it in getUpperBounds. I don't know if this matters, but it feels
consistent. Maybe it will save us from another surprising failure later. Or
maybe it will help reveal a case in which both type.getTypeVariable and
wildcardToTypeParam fail. We would probably like to hear about such cases.

I did verify that, even after I changed getUpperBounds, I do still need the
recently introduced special handling in visitWildcard. I had hoped that I might
not: *Hopefully* the change to getUpperBounds makes the checker see `Class<?
extends Object>` and `Class<?>` as compatible (though it's probably impossible
to tell, since the latter now gets rewritten!), so it may make visitWildcard
unnecessary in that case. But visitWildcard is still necessary in the case that
prompted it originally.