Revert "Hack to handle casting the result of a wildcard-returning method."
This reverts commit 9348e76115088149d55d9495a5c720a5f1d6cd34.

When testing against Guava, I discovered that that change produces a
StackOverflowError when code refers to `Enum<?>`.

I'm still not entirely sure why: The recursion goes through
NullSpecTypeAnnotator.visitWildcard and visitDeclared (which makes some
sense to me) but specifically through the super calls in both methods
(which makes less sense to me). I could have understood if the recusion
went through the new call to getAnnotatedType, but it does not. What's
especially strange is that we don't even modify the type that we pass to
the supertype significantly: We just attach a new annotation to it.
Perhaps something is being more eagerly initialized during the execution
of getAnnotatedType or isNullExclusiveUnderEveryParameterization?

Anyway, it's sad to undo the advance made by the original change, but
undoing it is a lesser evil than crashing.

Note that I have kept the part of subsequent commit
74f564f415046f6fc4e0200a7990a0db603db9c8 that changed getUpperBounds to
match the now-removed code in visitWildcard. If it causes us trouble, we
could revert it, too, since I'm not aware of any problem that it solved.