Add a very special-case hack for casting the result of Class.newInstance.
My previous attempts (detailed in
https://github.com/jspecify/checker-framework/issues/4#issuecomment-763111013)
failed, but this is a wart that I'd rather not people have to hit, so a
special case seems justified.

(Note that users could alternatively use
clazz.asSubclass(...).newInstance(), but again, they shouldn't have to.
And that's not necessarily even a 100% solution if they need to cast to
a generic type or even a type variable for which they don't have an
accompanying Class instance.)