Use `createAbstractValue` instead of `new CFValue`.
First, I thought I might need to subclass `CFValue`. Then, I thought I
might need to control which instance of `CFAnalysis` I pass to `CFValue`
(so that I can always pass the least-convenient-world version, which I
thought might help with `moreSpecificValue`). Either of those changes
could benefit from this commit, but it turns out I'm not going to do
either at the moment.

Still, I think this commit is The Right Thing to Do.

This isn't quite the same as
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/87b89f0443c0ebddcfdc8bdab9f054df4c5562b2,
but it reminded me of that.