If a value is equal to null, make it nullable, even if it was unspecified before.
As noted in
https://github.com/jspecify/jspecify/commit/8b44d536be24663ba25cddf4d1db908acadf8df2,
we probably should make it nullable even if it was *non-null* before.
I'll see if that's trivial. If not, I'll probably not worry about it: I
hope it's a rarer case this the unspecified case, and the unspecified
case already took me longer than I'd hoped to spend.