Make tests that use JSpecify samples pass.
I hope :)

In more detail:

- Use samples from the samples-google-prototype branch of the jspecify
repo, rather than the main branch. See
https://github.com/jspecify/jspecify/commit/c02287e9313f0652d27181e9aad69faec88cad0c

- Recognize the magic `jspecify_but_expect_*` comments from those
samples.

- Recognize a few new (or newly exercised) error keys. (Note that I
probably added some that aren't used by any of the released samples. For
example, I don't think that I released any `ThreadLocal`-related samples
(for reasons discussed in the commit referenced above), so I don't think
that `threadlocal.must.include.null` is needed yet. But it's harmless to
add now, and this will be the right place for it when the time comes.)

- Add `-AassumePure` to flags. (Also, remove a couple harmless but
unnecessary flags.)