Turn a couple unneeded methods into no-ops.
This shaves ~5 seconds off the time it takes to check Guava and ~1
second off the time it takes to check our samples.

(Rough numbers after this commit: 1:04 for my test script that checks
Guava, and :23 for my test script that checks the samples. Both of those
scripts spend a little time on other word (at least the Gradle
up-to-date check and, in the case of the samples, a google-java-format
run).)

(This commit salvages part of
https://github.com/jspecify/nullness-checker-for-checker-framework/pull/9)