Recognize `@CheckForNull`.
I thought I could just add this to my existing support for the protobuf
declaration annotations in `visitExecutable`. However, it turns out that that
support, while it worked for return types and for checking of arguments of
method calls, has never worked for other checks of parameters.

You can see the old samples for declaration annotations here:
https://github.com/jspecify/jspecify/blob/3714f6fa425e61cf3ab795a42be36b7af0d97151/samples/ProtoDeclarationAnnotations.java

Without digging too deeply, I'm reasonably confident that the underlying cause
is similar to that of another problem I've seen in the past. If so, the problem
here is that, if you ask CF for the type of a whole method, it will run
`visitExecutable`, as desired. And, when you ask for a method's return type or
parameter types, CF in fact asks for the type of the whole method and then
extracts what it needs. However, if you ask it only for the type of a parameter
"as a local variable," for example, then it presumably doesn't visit the whole
`ExecutableElement`, only that `VariableElement`.

The good news is that there doesn't seem to have been any reason for the
existing support for declaration annotations to use my custom logic. It would
be nice if I had written down *why* I felt that custom logic might be called
for. It's possible that I had fears around the changes I made to the handling
of declaration annotations in *stub* files for
https://github.com/jspecify/checker-framework/issues/11. But I have to wonder
if maybe I just didn't think that CF's built-in support would be as simple to
use as it turned out to be.

The newly expanded samples for declaration annotations pass when run as tests,
notably this one:
https://github.com/jspecify/jspecify/blob/4431f938f723c6ee406e35725ec82b50efaf6518/samples/UseCheckForNull.java#L34