Unsoundly assume that this.getClass().getClassLoader() never returns null.
This is similar to
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/880820f21706281c50b6cad246375c0f411f8006

This latest change is motivated by some code in Guava:
https://github.com/google/guava/blob/4732a867e2a986f09b28acee924ab74d92bd6ae7/guava/src/com/google/common/base/FinalizableReferenceQueue.java#L305

We could probably just change that to use
DecoupledLoader.class.getClassLoader() -- which would then be covered by the
previously added special case. But DecoupledLoader has an anonymous subclass in
Guava's tests, which perhaps theoretically could be loaded by a different class
loader (though that seems unlikely) and for all I know, there may be a good
reason for that (though that, too, seems unlikely). Really, though, we might as
well handle this in the checker so that other users can benefit.

I tested this against an expanded sample input for getClassLoader:
https://github.com/jspecify/jspecify/commit/ea558a3474dda3322ab8b02e0ef17a9a60250b5a