Recognize that path.getFileName() is non-null if the path comes from a DirectoryStream.
I *think* this is sound (assuming that no one implements DirectoryStream
in a perverse way). Handling this special case might not be worth the
trouble, but I already hit the case in both Turbine and Guava, so I
figured I might as well deal with it now.

Relevant sample input:
https://github.com/jspecify/jspecify/blob/44880e135d76366465be74d4a234d5797c513ac3/samples/DirectoryStreamGetFileName.java