Give up on making `Preconditions.checkNotNull` require a non-null input.
See the comment discussing this that we'll be adding to
`Preconditions.java` this week.

Part of the work will be done by changing how we annotate (or plan to
annotate) `Preconditions.java` itself. But our checker also has this
special case to explain why we were going to forbid nullable arguments.
So we no longer need that special case, and in fact we need to *not*
have it because it generates an error of its own.