Recognize that `Preconditions.checkNotNull` proves that an expression is non-null.
We'd already annotated the method to prove that its _return type_ is
non-null (though we haven't published a Maven release of Guava with that
change yet). This commit adds handling to recognize that a local
variable, field, etc. passed to `checkNotNull` is proven non-null in the
future. This makes us treat `checkNotNull` just like we already treat
`Objects.requireNonNull`.

(I can't point to a sample input that demonstrates this yet, but I
tested locally.)

Also, unrelated to all that, silence an Error Prone finding.