Move javax.lang.model instances into the newly instantiable Util class.
This puts all the initialization boilerplate in one place, including
eliminating some duplication. It also saves callers of methods like
hasSuppressWarningsNullness from having to pass in various objects from
the environment.

For similar reasons, move all AnnotatedTypeMirror instances into
NullSpecAnnotatedTypeFactory. Along the way, I created a createType
helper method.

Most code now refers to types, etc. through "util.foo." But in the case
of the AnnotationMirror instances for our nullness operators (unionNull,
etc.), I've been keeping the existing fields as shortcuts. That seems
worthwhile for the ability to say "somethingComplicated(unionNull, ...)"
over "somethingComplicated(util.unionNull, ...)."

The entry point for our checker is NullSpecChecker, so that's what
creates the Util instance. From there, we pass it to the constructor of
each of our other types. It's all straightforward except for the part
discussed in NullSpecChecker, which is actually more "about"
NullSpecVisitor.

The main voluntary change that I made along the way was to avoid
AnnotatedTypeMirror objects in favor of javax.lang.model types in a few
cases. This comes up in multiple classes, including a couple times in
NullSpecTransfer, but we'll look at NullSpecVisitor.visitNewClass:
Previously there, I had "javaLangThreadLocal.getUnderlyingType()."
Rather than change that to
"atypeFactory.javaLangThreadLocal.getUnderlyingType()," I chose to
change it to "util.javaLangThreadLocalElement.asType()." That's shorter,
and it sticks to javax.lang.model types.

Finally, here are 2 changes that I have thus far declined to make but
that I could be talked into:

- Renaming "Util" to something more like "Environment" (presumably
  renaming "util" fields to "env"), which would emphasize the connection
  with ProcessingEnvironment.

- Make the AnnotatedTypeMirror instances inside
  NullSpecAnnotatedTypeFactory private, exposing only methods like
  "asThreadLocal(AnnotatedTypeMirror)."