Prepare `AnnotatedTypeFormatter` for type variables produced by capture conversion.
I have some memory that such type variables *can* appear even in the
current version of the checker, but they aren't common: That version of
the checker is based on an upstream version that did not contain an
implementation of capture conversion for annotated types. Again, though,
I think they *might* still arise sometimes from direct usage of
javac-generated types, in which case this commit may improve error
messages in those cases. (Or maybe those cases are a lost cause.)

The real motivation for this change, though, is that we'll be merging
the upstream implementation of capture conversion into our checker. When
we do, we'll not only have to improve error messages for this newly
common kind of type variable, but we also have to track which type
variables we're currently visiting to avoid infinite recursion -- just
as we already do with wildcards.