Delay calling `postInit()` on either "world" for as long as we can.
I don't appear to have made a note about *specifically* what problem
this solved. But something in the upstream changes that I'm working to
merge required it: Something changed to make `postInit()` start trying
to use at least one of the
`withLeastConvenientWorld`/`withMostConvenientWorld` fields, probably
leading to NPE. This change is enough to avoid the problem.