Remove an older, inferior copy of a workaround I just almost exactly duplicated.
https://github.com/jspecify/nullness-checker-for-checker-framework/commit/824113a212a7ea001e6557fb800a5d73e8f18232
added almost the same check in a different location. The only difference
was that the new check included a `!isCaptured` check, as will be
required in a post-capture-conversion world.

It turns out that the `!isCaptured` check actually
[helps a little even now](https://github.com/jspecify/jspecify/commit/2a39064fe9b1daf281774121682369d2a7381a5d),
since the version of CF we're currently based on
[only _mostly_ doesn't encounter capture-conversion-created type variables](https://github.com/jspecify/nullness-checker-for-checker-framework/commit/e80571b1d612f2476f89b401fcd6100169c3f7c5).

And anyway, we don't want almost exactly the same code twice, especially
when the older version will soon start causing problems :)