Proof of concept for conformance tests
Create a mechanism to perform conformance tests. A conformance test analyzes a group of files and produces a report showing which assertions contained in those files pass or fail. The test can be run in three modes:

1. write the report to a file and pass
2. compare the report to the written file and pass if there are no changes
3. emit all failures

This also commits the file containing the report for the test source at https://github.com/jspecify/jspecify/tree/main/samples.