Simplify test report (#96)
Improve "details" mode output so that unexpected facts are reported
in-line with expected facts (with "OOPS" if the reported fact must be
expected, or "INFO" if not). This way it will be easier to see what is
going wrong when the comparison fails, and to see what other facts could
be expected.

Make "details" mode also run the comparison and fail, just with extra
output.

Also do some code simplification:
* `ConformanceTestReport` now holds all expected and reported facts and
their mapping, rather than a set of passing and failing assertions.
* Instead of a `ConformanceTestSubject` that does complex comparison
between the generated report and the stored one, use `Truth`'s very good
text comparison to compare the text of the generated report to the text
of the stored one.